Remove dependencies to GpuLower from runtime
