+++ csrc/device_lower/analysis/bank_conflict.cpp
@@ -7,6 +7,7 @@
 // clang-format on
 #include <device_lower/analysis/bank_conflict.h>
 
+#include <device_lower/utils.h>
 #include <expr_evaluator.h>
 #include <ir/utils.h>
 #include <kernel_ir.h>
