+++ csrc/device_lower/utils.h
@@ -15,7 +15,7 @@
 #include <ir/all_nodes.h>
 #include <kernel_ir.h>
 #include <parallel_type_bitmap.h>
-#include <val_graph.h>
+#include <val_graph_nodes.h>
 
 #include <bitset>
 #include <map>
@@ -25,6 +25,7 @@
 namespace nvfuser {
 
 class ThreadPredicateMap;
+class ValGraph;
 
 namespace scope_utils {
 
@@ -37,31 +38,6 @@ kir::IfThenElse* cloneIfThenElse(kir::IfThenElse* ite);
 } // namespace scope_utils
 
 namespace ir_utils {
-
-// Somtimes we want to temporarily view a tensorview with another tensordomain.
-// This isn't a permanent transformation, but in indexing we want to index
-// producers with a consumer set of indices, so we need to view the producer
-// transformed like consumer while we index. This will set the tv with td for
-// the life of this context guard.
-class TVDomainGuard {
- private:
-  TensorView* tv_;
-  TensorDomain* prev_domain_;
-
- public:
-  explicit TVDomainGuard(TensorView* tv, TensorDomain* td);
-  TVDomainGuard(const TVDomainGuard&) = delete;
-  NVF_API TVDomainGuard(TVDomainGuard&&);
-
-  //! An utility to access the tensordomain before the temporary
-  //!  view. This is used to retrieve information, like swizzle
-  //!  information that can only be reliably kept at the original domain.
-  const TensorDomain* prevDomain() const {
-    return prev_domain_;
-  }
-
-  NVF_API ~TVDomainGuard();
-};
 
 // Create a TVDomainGuard that temporarily view a TensorView with specified
 // all-true or all-false contiguity.
