+++ csrc/evaluator_common.h
@@ -6,7 +6,6 @@
  */
 // clang-format on
 #pragma once
-#include <device_lower/lower2device.h>
 #include <exceptions.h>
 #include <fusion.h>
 #include <ir/all_nodes.h>
