+++ csrc/runtime/executor.h
@@ -6,7 +6,6 @@
  */
 // clang-format on
 #pragma once
-#include <device_lower/lower2device.h>
 #include <exceptions.h>
 #include <expr_evaluator.h>
 #include <fusion.h>
@@ -85,6 +84,8 @@ struct KernelExecutorEntry {
   // requires an array of this form.
   std::vector<void*> arg_ptrs;
 };
+
+class GpuLower;
 
 class KernelExecutor : public ExecutorAbstract {
  public:
