+++ csrc/runtime/executor_utils.h
@@ -16,7 +16,6 @@
 #include <torch/csrc/jit/ir/ir.h>
 
 #include <cuda_utils.h>
-#include <device_lower/lower2device.h>
 #include <expr_evaluator.h>
 #include <fusion.h>
 #include <ir/all_nodes.h>
@@ -27,6 +26,9 @@
 #include <vector>
 
 namespace nvfuser {
+
+class GpuLower;
+
 namespace executor_utils {
 
 // I'm not happy with CudaExecutable being a struct exposing all the fields.
